Factor Axis, CrossAxisAlignment and MainAxisAlignment out of `flex.rs`
These types are going to be used in other widgets (notably the future Stack widget) and should not exclusive to Flex.
