+++ masonry/src/properties/types/mod.rs
@@ -12,8 +12,10 @@
 //! [`Background`]: crate::properties::Background
 //! [`Background::Gradient`]: crate::properties::Background::Gradient
 
+mod alignment;
 mod gradient;
 mod unit_point;
 
+pub use alignment::*;
 pub use gradient::*;
 pub use unit_point::UnitPoint;
