+++ xilem/examples/external_event_loop.rs
@@ -6,8 +6,8 @@
 //! accessing raw events from winit.
 //! Support for more custom embeddings would be welcome, but needs more design work
 
+use masonry::properties::types::{CrossAxisAlignment, MainAxisAlignment};
 use masonry::theme::default_property_set;
-use masonry::widgets::{CrossAxisAlignment, MainAxisAlignment};
 use masonry_winit::app::{AppDriver, MasonryUserEvent};
 use winit::application::ApplicationHandler;
 use winit::error::EventLoopError;
