+++ xilem/src/view/flex.rs
@@ -5,10 +5,11 @@ use std::marker::PhantomData;
 
 use crate::style::Style;
 
-use masonry::core::{FromDynWidget, Widget, WidgetMut};
+use masonry::core::{Axis, FromDynWidget, Widget, WidgetMut};
+pub use masonry::properties::types::{CrossAxisAlignment, MainAxisAlignment};
 use masonry::properties::{Background, BorderColor, BorderWidth, CornerRadius, Padding};
+pub use masonry::widgets::FlexParams;
 use masonry::widgets::{self};
-pub use masonry::widgets::{Axis, CrossAxisAlignment, FlexParams, MainAxisAlignment};
 
 use crate::core::{
     AppendVec, DynMessage, ElementSplice, MessageResult, Mut, SuperElement, View, ViewElement,
